Adding Constructors for deck and banner and board. linking score and timer
