+++ components/banner/banner.html
@@ -1,0 +1,7 @@
+<header id="banner">
+    <button class="logoutButton">↩️Logout</button>
+    <h1>Game of Spider</h1>
+    <p id="score">Score: 0</p>
+    <p id="timer">00:00:00</p>
+    <button class="button" id="retire">Retire</button>
+</header>
