+++ components/banner/banner.js
@@ -1,0 +1,16 @@
+const template = await fetch('/components/banner/banner.html')
+  .then(response => response.text())
+  .then(text => {
+    const template = document.createElement('template');
+    template.innerHTML = text;
+    return template;
+  });
+
+/**
+ * @returns {HTMLElement}
+ */
+export const createBanner = () => {
+    const bannerNode = template.content.cloneNode(true);
+    document.body.appendChild(bannerNode);
+    return document.getElementById('banner');
+  }
