+++ components/board/board.css
@@ -1,33 +1,4 @@
-.heading {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-}
-.title {
-    flex-grow: 1;
-    text-align: center;
-    font-size: 2rem;
-    font-family: 'Roboto', 'Arial Narrow', Arial, sans-serif;
-    margin: 0;
-}
-.logoutButton {
-    font-size: 1rem;
-    font-family: 'Roboto', 'Arial Narrow', Arial, sans-serif;
-    position: relative;
-    margin-left: 1rem;
-    cursor: pointer;
-}
-.time {
-    font-size: 1rem;
-    font-family: 'Roboto', 'Arial Narrow', Arial, sans-serif;
-    margin-right: 1rem;
-}
-section.heading {
-    position: relative;
-}
-.board {
+#board {
     gap: 1vw;
 
     &>* {
