+++ components/board/board.html
@@ -1,4 +1,4 @@
-<ul class='board container-h'>
+<ul id="board" class='container-h'>
     <li class="v-container">
         <button disabled class="empty"></button>
     </li>
