+++ components/deck/deck.html
@@ -1,0 +1,1 @@
+<button id="deck" class="container"></button>
