+++ index.html
@@ -11,6 +11,5 @@
     <title>Spoder Web</title>
   </head>
   <body class="v-container" onload="getJWT()">
-    <button id="deck" class="container"></button>
   </body>
 </html>
