+++ style/index.css
@@ -1,6 +1,7 @@
 @import "../components/card/card.css";
 @import "../components/board/board.css";
-@import "./deck.css";
+@import "../components/deck/deck.css";
+@import "../components/banner/banner.css";
 
 * {
     box-sizing: border-box;
@@ -16,6 +17,7 @@ button {
 }
 
 :root {
+    font-family: 'Roboto', 'Arial Narrow', Arial, sans-serif;
     padding: 1em;
     background-image: radial-gradient( circle at 100% -20%,  rgba(255,255,255,20%) 0%, rgba(255,255,255,0%) 70% );
     --color: black;
